# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DisableSkillRequest": {
        "properties": {
          "harness": {
            "description": "Harness identifier",
            "minLength": 1,
            "title": "Harness",
            "type": "string"
          }
        },
        "required": [
          "harness"
        ],
        "title": "DisableSkillRequest",
        "type": "object"
      },
...
      "EnableSkillRequest": {
        "properties": {
          "harness": {
            "description": "Harness identifier",
            "minLength": 1,
            "title": "Harness",
            "type": "string"
          }
        },
        "required": [
          "harness"
        ],
        "title": "EnableSkillRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HarnessCellResponse": {
        "properties": {
          "harness": {
            "title": "Harness",
            "type": "string"
          },
          "interactive": {
            "title": "Interactive",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "logoKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logokey"
          },
          "state": {
            "enum": [
              "enabled",
              "disabled",
              "found",
              "empty"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "harness",
          "label",
          "state",
          "interactive"
        ],
        "title": "HarnessCellResponse",
        "type": "object"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "type": "object"
      },
...
      "SetSkillHarnessesFailureResponse": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "harness": {
            "title": "Harness",
            "type": "string"
          }
        },
        "required": [
          "harness",
          "error"
        ],
        "title": "SetSkillHarnessesFailureResponse",
        "type": "object"
      },
      "SetSkillHarnessesRequest": {
        "properties": {
          "target": {
            "description": "Target state to apply to every interactive harness cell on this skill",
            "enum": [
              "enabled",
              "disabled"
            ],
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target"
        ],
        "title": "SetSkillHarnessesRequest",
        "type": "object"
      },
      "SetSkillHarnessesResultResponse": {
        "properties": {
          "failed": {
            "items": {
              "$ref": "#/components/schemas/SetSkillHarnessesFailureResponse"
            },
            "title": "Failed",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "succeeded": {
            "items": {
              "type": "string"
            },
            "title": "Succeeded",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "succeeded",
          "failed"
        ],
        "title": "SetSkillHarnessesResultResponse",
        "type": "object"
      },
...
      "SkillDetailActionsResponse": {
        "properties": {
          "canDelete": {
            "title": "Candelete",
            "type": "boolean"
          },
          "canManage": {
            "title": "Canmanage",
            "type": "boolean"
          },
          "deleteHarnessLabels": {
            "items": {
              "type": "string"
            },
            "title": "Deleteharnesslabels",
            "type": "array"
          },
          "stopManagingHarnessLabels": {
            "items": {
              "type": "string"
            },
            "title": "Stopmanagingharnesslabels",
            "type": "array"
          },
          "stopManagingStatus": {
            "anyOf": [
              {
                "enum": [
                  "available",
                  "disabled_no_enabled"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopmanagingstatus"
          }
        },
        "required": [
          "canManage",
          "stopManagingStatus",
          "stopManagingHarnessLabels",
          "canDelete",
          "deleteHarnessLabels"
        ],
        "title": "SkillDetailActionsResponse",
        "type": "object"
      },
      "SkillDetailResponse": {
        "properties": {
          "actions": {
            "$ref": "#/components/schemas/SkillDetailActionsResponse"
          },
          "attentionMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attentionmessage"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "displayStatus": {
            "enum": [
              "Managed",
              "Unmanaged"
            ],
            "title": "Displaystatus",
            "type": "string"
          },
          "documentMarkdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentmarkdown"
          },
          "harnessCells": {
            "items": {
              "$ref": "#/components/schemas/HarnessCellResponse"
            },
            "title": "Harnesscells",
            "type": "array"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/SkillLocationResponse"
            },
            "title": "Locations",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "skillRef": {
            "title": "Skillref",
            "type": "string"
          },
          "sourceLinks": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SkillSourceLinksResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "skillRef",
          "name",
          "description",
          "displayStatus",
          "attentionMessage",
          "actions",
          "harnessCells",
          "locations",
          "sourceLinks",
          "documentMarkdown"
        ],
        "title": "SkillDetailResponse",
        "type": "object"
      },
      "SkillLocationResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "harness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harness"
          },
          "kind": {
            "enum": [
              "shared",
              "harness"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "sourceKind": {
            "title": "Sourcekind",
            "type": "string"
          },
          "sourceLocator": {
            "title": "Sourcelocator",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "harness",
          "label",
          "scope",
          "path",
          "revision",
          "sourceKind",
          "sourceLocator",
          "detail"
        ],
        "title": "SkillLocationResponse",
        "type": "object"
      },
...
      "SkillSourceLinksResponse": {
        "properties": {
          "folderUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folderurl"
          },
          "repoLabel": {
            "title": "Repolabel",
            "type": "string"
          },
          "repoUrl": {
            "title": "Repourl",
            "type": "string"
          }
        },
        "required": [
          "repoLabel",
          "repoUrl",
          "folderUrl"
        ],
        "title": "SkillSourceLinksResponse",
        "type": "object"
      },
      "SkillSourceStatusResponse": {
        "properties": {
          "updateStatus": {
            "anyOf": [
              {
                "enum": [
                  "update_available",
                  "no_update_available",
                  "no_source_available",
                  "local_changes_detected"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatestatus"
          }
        },
        "required": [
          "updateStatus"
        ],
        "title": "SkillSourceStatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "skill-manager",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/skills/{skill_ref}": {
      "get": {
        "operationId": "get_skill_detail_api_skills__skill_ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Skill Detail"
      }
    },
    "/api/skills/{skill_ref}/delete": {
      "post": {
        "operationId": "delete_skill_api_skills__skill_ref__delete_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Skill"
      }
    },
    "/api/skills/{skill_ref}/disable": {
      "post": {
        "operationId": "disable_skill_api_skills__skill_ref__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableSkillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Skill"
      }
    },
    "/api/skills/{skill_ref}/enable": {
      "post": {
        "operationId": "enable_skill_api_skills__skill_ref__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableSkillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Skill"
      }
    },
    "/api/skills/{skill_ref}/manage": {
      "post": {
        "operationId": "manage_skill_api_skills__skill_ref__manage_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Manage Skill"
      }
    },
    "/api/skills/{skill_ref}/set-harnesses": {
      "post": {
        "operationId": "set_skill_harnesses_api_skills__skill_ref__set_harnesses_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSkillHarnessesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetSkillHarnessesResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Skill Harnesses"
      }
    },
    "/api/skills/{skill_ref}/source-status": {
      "get": {
        "operationId": "get_skill_source_status_api_skills__skill_ref__source_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillSourceStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Skill Source Status"
      }
    },
    "/api/skills/{skill_ref}/unmanage": {
      "post": {
        "operationId": "unmanage_skill_api_skills__skill_ref__unmanage_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unmanage Skill"
      }
    },
    "/api/skills/{skill_ref}/update": {
      "post": {
        "operationId": "update_skill_api_skills__skill_ref__update_post",
        "parameters": [
          {
            "in": "path",
            "name": "skill_ref",
            "required": true,
            "schema": {
              "title": "Skill Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Skill"
      }
    },
...
  }
}